pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "amccs"
version = "0.1.0"
description = "AMCCS - Android Multi-cam Camera System"
readme = "README.md"
authors = [{name = "Eusebiu A. Plesa"}]
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn>=0.23",
    "PyYAML>=6.0",
]

[project.urls]
Homepage = "https://github.com/eapCJ/amccs"
Repository = "https://github.com/eapCJ/amccs"
Issues = "https://github.com/eapCJ/amccs/issues"

[project.optional-dependencies]
test = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "httpx>=0.27",
]
dev = [
    "ruff>=0.6.5",
    "mypy>=1.9.0",
    "types-PyYAML>=6.0.12",
]

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
pythonpath = ["src"]
markers = [
    "integration: marks tests that hit real adb devices",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "ASYNC",
    "B",
    "BLE",
    "S",
    "PT",
    "RUF",
]
ignore = [
    "S311",  # tests intentionally use insecure randomness primitives
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = ["S101"]

[tool.mypy]
python_version = "3.11"
warn_unused_ignores = true
show_error_codes = true
pretty = true
